add auth and products controller
