+++ controllers/auth.js
@@ -1,1 +1,61 @@
+const router = require('express').Router()
+const User = require('../models/User')
+const CryptoJS = require('crypto-js')
+const jwt = require('jsonwebtoken')
 // register
+const registerUser = async (req, res) => {
+  const newUser = new User({
+    username: req.body.username,
+    email: req.body.email,
+    password: CryptoJS.AES.encrypt(
+      req.body.password,
+      process.env.PASS_SEC
+    ).toString(),
+  })
+
+  try {
+    const savedUser = await newUser.save()
+    res.status(201).json(savedUser)
+  } catch (err) {
+    res.status(500).json(err)
+  }
+}
+
+// login user
+
+const loginUser = async (req, res) => {
+  try {
+    const user = await User.findOne({
+      userName: req.body.user_name,
+    })
+
+    !user && res.status(401).json('Wrong User Name')
+
+    const hashedPassword = CryptoJS.AES.decrypt(
+      user.password,
+      process.env.PASS_SEC
+    )
+
+    const originalPassword = hashedPassword.toString(CryptoJS.enc.Utf8)
+
+    const inputPassword = req.body.password
+
+    originalPassword != inputPassword && res.status(401).json('Wrong Password')
+
+    const accessToken = jwt.sign(
+      {
+        id: user._id,
+        isAdmin: user.isAdmin,
+      },
+      process.env.JWT_SEC,
+      { expiresIn: '3d' }
+    )
+
+    const { password, ...others } = user._doc
+    res.status(200).json({ ...others, accessToken })
+  } catch (err) {
+    res.status(500).json(err)
+  }
+}
+
+module.exports = { registerUser, loginUser }
