+++ routes/cart.js
@@ -1,27 +1,27 @@
-const Cart = require("../models/Cart");
+const Cart = require('../models/Cart')
 const {
   verifyToken,
   verifyTokenAndAuthorization,
   verifyTokenAndAdmin,
-} = require("./verifyToken");
+} = require('./verifyToken')
 
-const router = require("express").Router();
+const router = require('express').Router()
 
 //CREATE
 
-router.post("/", verifyToken, async (req, res) => {
-  const newCart = new Cart(req.body);
+router.post('/', verifyToken, async (req, res) => {
+  const newCart = new Cart(req.body)
 
   try {
-    const savedCart = await newCart.save();
-    res.status(200).json(savedCart);
+    const savedCart = await newCart.save()
+    res.status(200).json(savedCart)
   } catch (err) {
-    res.status(500).json(err);
+    res.status(500).json(err)
   }
-});
+})
 
 //UPDATE
-router.put("/:id", verifyTokenAndAuthorization, async (req, res) => {
+router.put('/:id', verifyTokenAndAuthorization, async (req, res) => {
   try {
     const updatedCart = await Cart.findByIdAndUpdate(
       req.params.id,
@@ -29,42 +29,42 @@ router.put("/:id", verifyTokenAndAuthorization, async (req, res) => {
         $set: req.body,
       },
       { new: true }
-    );
-    res.status(200).json(updatedCart);
+    )
+    res.status(200).json(updatedCart)
   } catch (err) {
-    res.status(500).json(err);
+    res.status(500).json(err)
   }
-});
+})
 
 //DELETE
-router.delete("/:id", verifyTokenAndAuthorization, async (req, res) => {
+router.delete('/:id', verifyTokenAndAuthorization, async (req, res) => {
   try {
-    await Cart.findByIdAndDelete(req.params.id);
-    res.status(200).json("Cart has been deleted...");
+    await Cart.findByIdAndDelete(req.params.id)
+    res.status(200).json('Cart has been deleted...')
   } catch (err) {
-    res.status(500).json(err);
+    res.status(500).json(err)
   }
-});
+})
 
 //GET USER CART
-router.get("/find/:userId", verifyTokenAndAuthorization, async (req, res) => {
+router.get('/find/:userId', verifyTokenAndAuthorization, async (req, res) => {
   try {
-    const cart = await Cart.findOne({ userId: req.params.userId });
-    res.status(200).json(cart);
+    const cart = await Cart.findOne({ userId: req.params.userId })
+    res.status(200).json(cart)
   } catch (err) {
-    res.status(500).json(err);
+    res.status(500).json(err)
   }
-});
+})
 
 // //GET ALL
 
-router.get("/", verifyTokenAndAdmin, async (req, res) => {
+router.get('/', verifyTokenAndAdmin, async (req, res) => {
   try {
-    const carts = await Cart.find();
-    res.status(200).json(carts);
+    const carts = await Cart.find()
+    res.status(200).json(carts)
   } catch (err) {
-    res.status(500).json(err);
+    res.status(500).json(err)
   }
-});
+})
 
-module.exports = router;
+module.exports = router
